requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.0
cron-descriptor==1.4.5
dj-database-url==0.5.0
Django==5.1.6
django-appconf==1.1.0
django-autoslug==1.9.9
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-imagekit==5.0.0
django-soft-delete==1.0.18
django-softdelete==0.11.5
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-standardized-errors==0.14.1
drf-yasg==1.21.8
flower==2.0.1
gunicorn==23.0.0
h11==0.16.0
humanize==4.12.1
idna==3.10
inflection==0.5.1
kombu==5.4.2
Markdown==3.7
packaging==24.2
pilkit==3.0
pillow==11.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tornado==6.4.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
